+++ src/screens/Camera/CameraPage.tsx
@@ -1,33 +1,26 @@
-import React, {useState, useRef} from 'react';
-import {Text, View, Button} from 'react-native';
-import {CameraView, CameraType, useCameraPermissions} from 'expo-camera';
-import {useIsFocused} from '@react-navigation/native';
-import {launchImageLibrary} from 'react-native-image-picker';
-import {useBottomTabBarHeight} from '@react-navigation/bottom-tabs';
+import {CameraView} from 'expo-camera';
+import React from 'react';
+import {Button, Text, View} from 'react-native';
 
+import {ImagePickerButton, TakePictureButton} from '@/components/Buttons';
 import FocusAwareStatusBar from '@/components/FocusAwareStatusBar';
-import {TakePictureButton, ImagePickerButton} from '@/components/Buttons';
-
-import ResizeImage from '@/utils/resizeImage';
-
-import styles from './CameraPage.styles';
 import {DarkTheme} from '@/styles/theme';
-
-import {PhotoResizeResult} from '@/types/photo';
 import {CameraPageProps} from '@/types/navigation';
 
-const CameraPage = ({navigation}: CameraPageProps) => {
-  const cameraRef = useRef<CameraView>(null);
+import styles from '@/screens/Camera/CameraPage.styles';
+import {useCamera} from '@/screens/Camera/hooks/useCamera';
 
-  const bottomTabHeight = useBottomTabBarHeight();
-
-  // Camera states
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [facing, setFacing] = useState<CameraType>('back');
-  const [permission, requestPermission] = useCameraPermissions();
-
-  // Only show the camera when the tab is focused
-  const isFocused = useIsFocused();
+const CameraPage = ({navigation, route}: CameraPageProps) => {
+  const {
+    cameraRef,
+    bottomTabHeight,
+    facing,
+    permission,
+    requestPermission,
+    isFocused,
+    takePicture,
+    pickImage,
+  } = useCamera({navigation, route});
 
   // Permission logic
 
@@ -45,31 +38,6 @@ const CameraPage = ({navigation}: CameraPageProps) => {
       </View>
     );
   }
-
-  const takePicture = async () => {
-    if (!cameraRef.current) {
-      return;
-    }
-    const photo = await cameraRef.current.takePictureAsync({
-      shutterSound: false,
-    });
-    const resizedPhoto = await ResizeImage(photo.uri);
-    navigateToDetect(resizedPhoto);
-  };
-
-  const pickImage = async () => {
-    const photo = await launchImageLibrary({mediaType: 'photo'});
-    if (photo.didCancel === true) {
-      return;
-    }
-    const resizedPhoto = await ResizeImage(photo.assets[0].uri);
-    navigateToDetect(resizedPhoto);
-  };
-
-  // Navigation logic
-  const navigateToDetect = (photo: PhotoResizeResult) => {
-    navigation.navigate('detect-page', {photo});
-  };
 
   return (
     <View style={styles.container}>
