+++ src/screens/Camera/ImageDetectPage.tsx
@@ -1,37 +1,20 @@
-// Lib imports
-import React, {useRef, useEffect, useState} from 'react';
-import {View, ImageBackground, Linking} from 'react-native';
-import {useHeaderHeight} from '@react-navigation/elements';
-import {useBottomTabBarHeight} from '@react-navigation/bottom-tabs';
-import {useSafeAreaInsets} from 'react-native-safe-area-context';
+import {ArrowLeft, Refresh, Search} from 'iconoir-react-native';
+import React from 'react';
+import {ImageBackground, View} from 'react-native';
 
-// Icon imports
-import {Search} from 'iconoir-react-native';
-import {Refresh} from 'iconoir-react-native';
-import {ArrowLeft} from 'iconoir-react-native';
-
-// Component imports
-import {LoadingIndicator} from '@/components/LoadingIndicator';
-import FocusAwareStatusBar from '@/components/FocusAwareStatusBar';
-import {ErrorChip} from '@/components/ErrorMessage';
-import {GoButton} from '@/components/Buttons';
 import {DarkPersistentActionSheet} from '@/components/ActionSheet/';
-
-// Helper imports
+import {GoButton} from '@/components/Buttons';
+import {ErrorChip} from '@/components/ErrorMessage';
+import FocusAwareStatusBar from '@/components/FocusAwareStatusBar';
+import {LoadingIndicator} from '@/components/LoadingIndicator';
 import {SelectedResultContext} from '@/contexts/DetectionResultContext';
-import {DetectResultRenderer} from './components/DetectResultRenderer';
-import {useDetection} from './hooks/useDetection';
-import {useActionSheetInitPoint} from '@/hooks/useActionSheetInitPoint';
-
-// Style imports
-import {styles} from './ImageDetectPage.styles';
-// This page ALWAYS use dark theme
 import {DarkTheme} from '@/styles/theme';
-
-// Type imports
 import {DetectionResultType} from '@/types/detection';
 import {ImageDetectPageProps} from '@/types/navigation';
-import {ActionSheetRef} from 'react-native-actions-sheet';
+
+import {DetectResultRenderer} from '@/screens/Camera/components/DetectResultRenderer';
+import {useImageDetect} from '@/screens/Camera/hooks/useImageDetect';
+import {styles} from '@/screens/Camera/ImageDetectPage.styles';
 
 type DetectResultProps = {
   fetchResult: DetectionResultType[];
@@ -56,65 +39,23 @@ const DetectResult = ({fetchResult, type}: DetectResultProps) => {
 };
 
 const ImageDetectPage = ({route, navigation}: ImageDetectPageProps) => {
-  // Get passed photo from routes
-  const {photo} = route.params;
-
-  // RN navigation
-  const headerHeight = useHeaderHeight();
-  const bottomTabHeight = useBottomTabBarHeight();
-
-  // Inset
-  const insets = useSafeAreaInsets();
-
-  // Result selecting state
-  const [selectedResult, setSelectedResult] = useState({
-    result: '',
-    index: -1,
-  });
-
-  // Image resizing states
-  const [resizeRatio, setResizeRatio] = useState(1);
-  const [imageWidthDevice, setImageWidthDevice] = useState(0);
-
-  // Detection results states and function
-  const {detectionState, getData} = useDetection();
-
-  // Action sheet logic
-  const resultsActionSheetRef = useRef<ActionSheetRef>(null);
-  // Ref: @/components/ActionSheet/ActionSheet.styles.ts
-  const sheetIndicatorHeight = DarkTheme.spacing.md * 2 + DarkTheme.spacing.sm;
-  const {setSheetChildrenHeight, initSnapPoint} = useActionSheetInitPoint(
-    // This AS automatically adds the OS navbar height to its height
-    // To calculate the desired location, we need to minus it
-    sheetIndicatorHeight + bottomTabHeight - insets.bottom,
-  );
-
-  function openActionSheet() {
-    resultsActionSheetRef.current?.show();
-  }
-
-  useEffect(() => {
-    if (!detectionState.isLoading) {
-      openActionSheet();
-    }
-  }, [detectionState.isLoading]);
-
-  // Fetch data on first render
-
-  useEffect(() => {
-    getData(photo.uri);
-  }, [photo, getData]);
-
-  // Navigation logic
-  function goBack() {
-    navigation.navigate('main-camera');
-  }
-
-  function searchMap() {
-    Linking.openURL(
-      'http://maps.google.com/?q=' + selectedResult.result + ' shop',
-    );
-  }
+  const {
+    photo,
+    setResizeRatio,
+    setImageWidthDevice,
+    detectionState,
+    imageWidthDevice,
+    resizeRatio,
+    selectedResult,
+    setSelectedResult,
+    headerHeight,
+    resultsActionSheetRef,
+    initSnapPoint,
+    setSheetChildrenHeight,
+    goBack,
+    getData,
+    searchMap,
+  } = useImageDetect({route, navigation});
 
   return (
     <View style={styles.container}>
