+++ src/screens/Camera/navigator.tsx
