--- src/screens/HistoryPage.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const HistoryPage = () => {
-  return <></>;
-};
-
-export default HistoryPage;
