+++ src/screens/User/UserPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const UserPage = () => {
+  return <></>;
+};
+
+export default UserPage;
